refactor recruitment views: move to separate module and update router registrations
